Corrected authenticatable
The login method was causing a infinite loop to be created when a user was authenticated? but not a admin. This would send them into a loop always being redirected back to the cas login. Since they were already authenticated this caused too many redirects since the cas page already detected a valid login.